+++ fabric/src/main/java/io/github/sakurawald/module/initializer/chat/mention/ChatMentionInitializer.java
@@ -8,6 +8,7 @@ import io.github.sakurawald.core.auxiliary.minecraft.TextHelper;
 import io.github.sakurawald.core.config.handler.abst.BaseConfigurationHandler;
 import io.github.sakurawald.core.config.handler.impl.ObjectConfigurationHandler;
 import io.github.sakurawald.core.job.impl.MentionPlayersJob;
+import io.github.sakurawald.core.structure.descriptor.annotation.ColorBox;
 import io.github.sakurawald.module.initializer.ModuleInitializer;
 import io.github.sakurawald.module.initializer.chat.mention.config.model.ChatMentionConfigModel;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -23,6 +24,14 @@ import java.util.List;
     1. The target player name will be highlighted.
     2. The target player will be sound notified.
     """)
+
+@ColorBox(color = ColorBox.ColorBlockTypes.WARNING, value = """
+    This module work partially with `Styled Chat` mod.
+    You can use this module with that mod.
+    It works, but you will not get the `mention player` rendered in chat.
+    Other things like the `sound notify` will still work.
+    """)
+
 public class ChatMentionInitializer extends ModuleInitializer {
     private static final BaseConfigurationHandler<ChatMentionConfigModel> config = new ObjectConfigurationHandler<>(BaseConfigurationHandler.CONFIG_JSON, ChatMentionConfigModel.class);