+++ fabric/src/main/java/io/github/sakurawald/module/initializer/chat/replace/ChatReplaceInitializer.java
@@ -6,6 +6,7 @@ import io.github.sakurawald.core.auxiliary.minecraft.TextHelper;
 import io.github.sakurawald.core.config.handler.abst.BaseConfigurationHandler;
 import io.github.sakurawald.core.config.handler.impl.ObjectConfigurationHandler;
 import io.github.sakurawald.core.structure.RegexRewriteNode;
+import io.github.sakurawald.core.structure.descriptor.annotation.ColorBox;
 import io.github.sakurawald.module.initializer.ModuleInitializer;
 import io.github.sakurawald.module.initializer.chat.replace.model.ChatReplaceConfigModel;
 import net.minecraft.entity.player.PlayerEntity;
@@ -15,6 +16,15 @@ import net.minecraft.text.Text;
 @Document("""
     This module allows you to replace `chat string` with a given `chat text`. (Including placeholder parsing)
     """)
+
+@ColorBox(color = ColorBox.ColorBlockTypes.WARNING, value = """
+    This module didn't work with `Styled Chat` mod.
+    The `chat.replace` module does work with `chat.style` module.
+    But if you are using the `Styled Chat` mod, then it didn't work.
+    However, since `Styled Chat` mod allows you to define custom `emotions`.
+    So you can `disable` this module, if you are using that mod.
+    """)
+
 public class ChatReplaceInitializer extends ModuleInitializer {
 
     private static final BaseConfigurationHandler<ChatReplaceConfigModel> config = new ObjectConfigurationHandler<>(BaseConfigurationHandler.CONFIG_JSON, ChatReplaceConfigModel.class);