+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **This version is the `Quality of Life` improvement version.**
 
+- [fuji] feature: add more the `colorboxes` for `all modules`. (Provides more document strings.)
 - [core] feature: provide better warning messages when failed to initialize a module on server startup.
 - [core] feature: use a better implementation to invoke module initializers, and skip the initialization of variables if the module is disabled.
 - [fuji] fix: should escape the `placeholders` and `style tags` when displaying `document string`.